# openapi.yaml
openapi: 3.1.0
info:
  title: PEG HR LMS v1 API
  version: "1.0.0"
  description: >
    HTTP-контракт системы онбординга новых сотрудников (см. SPEC.md).
    Один сервис. Все пути под /api/v1. Реализовано в prototype/.
    Ошибка любого запроса: { error: { code, message, details? }, request_id }.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Err:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "snake_case машинный код" }
            message: { type: string }
            details: {}
...
security:
  - bearer: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Вход по логину и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  role: { type: string, enum: [employee, hr, admin] }
        "401": { $ref: '#/components/responses/Err' }
...